custom/permissions.py
from rest_framework.permissions import BasePermission


# Set Permission for Super User [Super Admin]
class isSuperuser(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.is_superuser)


# Set Permission for Admin
class isAdmin(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.is_adminuser)


# Set Permission for Customer
class IsCustomer(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.is_customer)
