requirements.txt
asgiref==3.6.0
autopep8==2.0.1
cffi==1.15.1
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.14.0
git+https://github.com/jimmypappachan/django-models-logging@v3#egg=django_models_logging
django-rest-knox==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
ua-parser==0.16.1
user-agents==2.2.0